Add cubit for fetching balances and add helper widgets
