+++ lib/screens/friends_screen/widgets/tabs/my_friends_tab.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:kasa_w_grupie/cubits/friends_cubit.dart';
-import 'package:kasa_w_grupie/screens/friends_details_screen.dart';
+import 'package:kasa_w_grupie/cubits/group_balance_cubit.dart';
+import 'package:kasa_w_grupie/screens/friends_details_screen/friends_details_screen.dart';
 import 'package:kasa_w_grupie/services/friends_service.dart';
 
 class MyFriendsTab extends StatelessWidget {
@@ -10,11 +11,11 @@ class MyFriendsTab extends StatelessWidget {
   const MyFriendsTab({super.key, required this.friendsService});
 
   // Helper function to get balance info
-  String getBalanceInfo(double owesAmount, double owedAmount) {
-    if (owesAmount > 0) {
-      return "Owes you: ${owesAmount.toStringAsFixed(2)}";
-    } else if (owedAmount > 0) {
-      return "You owe: ${owedAmount.toStringAsFixed(2)}";
+  String getBalanceInfo(double totalBalance) {
+    if (totalBalance > 0) {
+      return "Owes you: ${totalBalance.toStringAsFixed(2)}";
+    } else if (totalBalance < 0) {
+      return "You owe: ${totalBalance.abs().toStringAsFixed(2)}";
     } else {
       return "No debts";
     }
@@ -36,46 +37,41 @@ class MyFriendsTab extends StatelessWidget {
             itemBuilder: (context, index) {
               final friend = state.friends[index];
 
-              return FutureBuilder<Map<String, dynamic>>(
-                future: friendsService.getUserBalances(friend.id),
-                builder: (context, snapshot) {
-                  String balanceInfo = "Loading...";
+              return BlocProvider(
+                create: (context) => GroupBalanceCubit(
+                  friendsService: friendsService,
+                )..loadGroupBalances(friend.id),
+                child: BlocBuilder<GroupBalanceCubit, GroupBalanceState>(
+                  builder: (context, balanceState) {
+                    String balanceInfo = "Loading...";
 
-                  if (snapshot.connectionState == ConnectionState.done) {
-                    if (snapshot.hasData) {
-                      final data = snapshot.data!;
-                      final bool? isOwedToUser = data["isOwedToUser"];
-                      final double amount = data["amount"] ?? 0.0;
-
-                      balanceInfo = getBalanceInfo(
-                        isOwedToUser == true ? amount : 0.0,
-                        isOwedToUser == false ? amount : 0.0,
-                      );
-                    } else {
+                    if (balanceState is GroupBalanceLoaded) {
+                      balanceInfo = getBalanceInfo(balanceState.totalBalance);
+                    } else if (balanceState is GroupBalanceError) {
                       balanceInfo = "Error loading balance";
                     }
-                  }
 
-                  return ListTile(
-                    leading: CircleAvatar(child: Icon(Icons.person)),
-                    title: Text(friend.name),
-                    subtitle: Text(balanceInfo),
-                    trailing: Icon(Icons.arrow_forward_ios),
-                    onTap: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                          builder: (context) => FriendDetailsScreen(
-                            userId: friend.id,
-                            friendName: friend.name,
-                            friendEmail: friend.email,
-                            friendsService: friendsService,
+                    return ListTile(
+                      leading: CircleAvatar(child: Icon(Icons.person)),
+                      title: Text(friend.name),
+                      subtitle: Text(balanceInfo),
+                      trailing: Icon(Icons.arrow_forward_ios),
+                      onTap: () {
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                            builder: (context) => FriendDetailsScreen(
+                              userId: friend.id,
+                              friendName: friend.name,
+                              friendEmail: friend.email,
+                              friendsService: friendsService,
+                            ),
                           ),
-                        ),
-                      );
-                    },
-                  );
-                },
+                        );
+                      },
+                    );
+                  },
+                ),
               );
             },
           );
